[IMP] Update model name and fields
