+++ base_model_restrict_update/README.rst
@@ -14,7 +14,7 @@ Update Restrict Model
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-qrtl%2Fhls--custom-lightgray.png?logo=github
-    :target: https://github.com/qrtl/hls-custom/tree/12.0/base_update_restrict_model
+    :target: https://github.com/qrtl/hls-custom/tree/12.0/base_model_restrict_update
     :alt: qrtl/hls-custom
 
 |badge1| |badge2| |badge3| 
@@ -39,7 +39,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/qrtl/hls-custom/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/qrtl/hls-custom/issues/new?body=module:%20base_update_restrict_model%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/qrtl/hls-custom/issues/new?body=module:%20base_model_restrict_update%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -54,6 +54,6 @@ Authors
 Maintainers
 ~~~~~~~~~~~
 
-This module is part of the `qrtl/hls-custom <https://github.com/qrtl/hls-custom/tree/12.0/base_update_restrict_model>`_ project on GitHub.
+This module is part of the `qrtl/hls-custom <https://github.com/qrtl/hls-custom/tree/12.0/base_model_restrict_update>`_ project on GitHub.
 
 You are welcome to contribute.
