+++ base_model_restrict_update/i18n/base_model_restrict_update.po
@@ -1,0 +1,56 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* base_model_restrict_update
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: base_model_restrict_update
+#: model:ir.model,name:base_model_restrict_update.model_ir_model_access
+msgid "Model Access"
+msgstr ""
+
+#. module: base_model_restrict_update
+#: model:ir.model,name:base_model_restrict_update.model_ir_model
+msgid "Models"
+msgstr ""
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,help:base_model_restrict_update.field_res_users__unrestrict_model_update
+msgid "Set to true and the user can update restricted model."
+msgstr ""
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,field_description:base_model_restrict_update.field_res_users__unrestrict_model_update
+msgid "Unrestrict Model Update"
+msgstr ""
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,field_description:base_model_restrict_update.field_ir_model__restrict_update
+msgid "Update Restrict Model"
+msgstr ""
+
+#. module: base_model_restrict_update
+#: model:ir.model,name:base_model_restrict_update.model_res_users
+msgid "Users"
+msgstr ""
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,help:base_model_restrict_update.field_ir_model__restrict_update
+msgid "When selected, the model is restricted to read-only unless the user has the special permission."
+msgstr ""
+
+#. module: base_model_restrict_update
+#: code:addons/base_model_restrict_update/models/ir_model_access.py:31
+#, python-format
+msgid "You are only allowed to read this record. (%s - %s)"
+msgstr ""
+
