+++ base_model_restrict_update/i18n/ja_JP.po
@@ -1,0 +1,58 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* base_model_restrict_update
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2021-02-04 09:47+0000\n"
+"PO-Revision-Date: 2021-02-04 09:47+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: base_model_restrict_update
+#: model:ir.model,name:base_model_restrict_update.model_ir_model_access
+msgid "Model Access"
+msgstr "モデルアクセス"
+
+#. module: base_model_restrict_update
+#: model:ir.model,name:base_model_restrict_update.model_ir_model
+msgid "Models"
+msgstr "モデル"
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,help:base_model_restrict_update.field_res_users__unrestrict_model_update
+msgid "Set to true and the user can update restricted model."
+msgstr "有効化にすると制限されたモデルを更新することができます。"
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,field_description:base_model_restrict_update.field_res_users__unrestrict_model_update
+msgid "Unrestrict Model Update"
+msgstr "制限モデル更新許可"
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,field_description:base_model_restrict_update.field_ir_model__restrict_update
+msgid "Update Restrict Model"
+msgstr "更新制限モデル"
+
+#. module: base_model_restrict_update
+#: model:ir.model,name:base_model_restrict_update.model_res_users
+msgid "Users"
+msgstr "ユーザ"
+
+#. module: base_model_restrict_update
+#: model:ir.model.fields,help:base_model_restrict_update.field_ir_model__restrict_update
+msgid "When selected, the model is restricted to read-only unless the user has the special permission."
+msgstr "有効化にすると、「制限モデル更新許可」を持ちユーザのみモデルレコードを更新できます。"
+
+#. module: base_model_restrict_update
+#: code:addons/base_model_restrict_update/models/ir_model_access.py:31
+#, python-format
+msgid "You are only allowed to read this record. (%s - %s)"
+msgstr "レコードを更新することはできません。(%s - %s)"
+
