+++ base_model_restrict_update/static/description/index.html
@@ -367,7 +367,7 @@ ul.auto-toc {
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/lgpl-3.0-standalone.html"><img alt="License: LGPL-3" src="https://img.shields.io/badge/licence-LGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/qrtl/hls-custom/tree/12.0/base_update_restrict_model"><img alt="qrtl/hls-custom" src="https://img.shields.io/badge/github-qrtl%2Fhls--custom-lightgray.png?logo=github" /></a></p>
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/lgpl-3.0-standalone.html"><img alt="License: LGPL-3" src="https://img.shields.io/badge/licence-LGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/qrtl/hls-custom/tree/12.0/base_model_restrict_update"><img alt="qrtl/hls-custom" src="https://img.shields.io/badge/github-qrtl%2Fhls--custom-lightgray.png?logo=github" /></a></p>
 <p>This module adds a config to apply a global update restriction to specific model,
 while only certain users can update the records if the config is enabled.</p>
 <p><strong>Table of contents</strong></p>
@@ -392,7 +392,7 @@ To set an user as a permitted user to update restricted model(s), click on “Gr
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/qrtl/hls-custom/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/qrtl/hls-custom/issues/new?body=module:%20base_update_restrict_model%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/qrtl/hls-custom/issues/new?body=module:%20base_model_restrict_update%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -405,7 +405,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 </div>
 <div class="section" id="maintainers">
 <h2><a class="toc-backref" href="#id5">Maintainers</a></h2>
-<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/hls-custom/tree/12.0/base_update_restrict_model">qrtl/hls-custom</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/hls-custom/tree/12.0/base_model_restrict_update">qrtl/hls-custom</a> project on GitHub.</p>
 <p>You are welcome to contribute.</p>
 </div>
 </div>
