+++ base_model_restrict_update/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_base_model_restrict_update
