--- base_update_restrict_model/i18n/base_update_restrict_model.po
@@ -1,51 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* base_update_restrict_model
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: base_update_restrict_model
-#: model:ir.model,name:base_update_restrict_model.model_ir_model_access
-msgid "Model Access"
-msgstr ""
-
-#. module: base_update_restrict_model
-#: model:ir.model,name:base_update_restrict_model.model_ir_model
-msgid "Models"
-msgstr ""
-
-#. module: base_update_restrict_model
-#: model:ir.model.fields,help:base_update_restrict_model.field_res_users__restrict_update_permit
-msgid "Set True Means This User can update restricted model."
-msgstr ""
-
-#. module: base_update_restrict_model
-#: model:ir.model.fields,field_description:base_update_restrict_model.field_res_users__restrict_update_permit
-msgid "Update Permit"
-msgstr ""
-
-#. module: base_update_restrict_model
-#: model:ir.model.fields,field_description:base_update_restrict_model.field_ir_model__update_restrict
-msgid "Update Restrict Model"
-msgstr ""
-
-#. module: base_update_restrict_model
-#: model:ir.model,name:base_update_restrict_model.model_res_users
-msgid "Users"
-msgstr ""
-
-#. module: base_update_restrict_model
-#: code:addons/base_update_restrict_model/models/ir_model_access.py:31
-#, python-format
-msgid "You are only allowed to read this record. (%s - %s)"
-msgstr ""
-
