--- base_update_restrict_model/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_base_update_restrict_model
